Move Smart Agent refactored code to separate package
