+++ receiver/k8sclusterreceiver/collection/metadatastore.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8sclusterreceiver
+package collection
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
