+++ receiver/k8sclusterreceiver/testutils/mock_cache_store.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package k8sclusterreceiver
+package testutils
 
 import "k8s.io/client-go/tools/cache"
